node-main/test/fixtures/wpt/encoding/remove-only-one-bom.html

<!doctype html>
<meta charset=utf-8>
<title></title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script>
setup({explicit_done:true});
function runTests() {
    for (var i = 0; i < window.frames.length; i++) {
        test(function() {
            assert_equals(window.frames[i].window.document.body.textContent, "\uFEFF");
        }, "Should have removed only one BOM from frame " + i);
    }
    done();
}
</script>
<body onload="runTests()">
<iframe src="resources/two-boms-utf-8.html"></iframe>
<iframe src="resources/two-boms-utf-16le.html"></iframe>
<iframe src="resources/two-boms-utf-16be.html"></iframe>
