third_party/WebKit/LayoutTests/external/wpt/encoding/remove-only-one-bom-expected.txt

This is a testharness.js-based test.
PASS Should have removed only one BOM from frame 0
FAIL Should have removed only one BOM from frame 1 assert_equals: expected "﻿" but got "\ufffd\ufffd\ufffd\ufffd\ufffd\ufffd\ufffd\ufffd\ufffd\ufffd\ufffd\ufffd"
FAIL Should have removed only one BOM from frame 2 assert_equals: expected "﻿" but got "\ufffd\ufffd\ufffd\ufffd\ufffd\ufffd\ufffd\ufffd\ufffd\ufffd\ufffd\ufffd"
Harness: the test ran to completion.
